better did-you-mean
ensure 'node' => 'nodejs' insteadof 'node-server'
